Move to new location in tests
